datenbank online

+++ client-app/src/views/Logout.vue
@@ -1,14 +1,13 @@
 <template>
-  <div class="container">
+  <div class="container-fluid px-3 px-md-5 px-lg-3 px-xl-5 py-5 mx-auto">
     <div class="d-flex flex-column justify-content-center">
-      <div class="mx-auto">
+      <div class="text-center mt-5">
         <router-link to="/">Home</router-link> |
         <router-link to="/login">Login</router-link>
       </div>
-      <h5 class="text-center my-3 text-danger">
-        Your are logged out!
-      </h5>
-      <p class="text-center mt-3">Bye!</p>
+      <div class="mb-5" style="text-align: center">
+        <h1 class="mt-5">You are not logged in!</h1>
+      </div>
     </div>
   </div>
 </template>
@@ -16,10 +15,16 @@
 <script>
 import axios from 'axios';
 export default {
+  data() {
+    return {};
+  },
   async created() {
-    localStorage.clear();
-    this.$router.push('/login');
-    await axios.get('/logout');
+    let user = JSON.parse(localStorage.getItem('user'));
+    if (user != null) {
+      localStorage.clear();
+      this.$router.push('/login');
+      await axios.get('/logout');
+    }
   },
 };
 </script>
